jenkins: Check if the node exists before deleting
Otherwise, we run into problems like:

$ jcs delete --cloud openstack sesdev-integration_PR-351-8
Connected to jenkins server 2.190.2
Jenkins node "sesdev-integration_PR-351-8" deleting...
Traceback (most recent call last):
  File "/tmp/ws/sesdev-integration/PR-351/8/jcs/venv/bin/jcs", line 8, in <module>
    sys.exit(main())
  File "/tmp/ws/sesdev-integration/PR-351/8/jcs/venv/lib/python3.6/site-packages/jcs/__init__.py", line 285, in main
    args.func(args)
  File "/tmp/ws/sesdev-integration/PR-351/8/jcs/venv/lib/python3.6/site-packages/jcs/__init__.py", line 202, in _do_delete
    jen_client.delete_node(args.jenkins_name)
  File "/tmp/ws/sesdev-integration/PR-351/8/jcs/venv/lib/python3.6/site-packages/jcs/jen.py", line 36, in delete_node
    self._client.delete_node(node_name)
  File "/tmp/ws/sesdev-integration/PR-351/8/jcs/venv/lib/python3.6/site-packages/jenkinsapi/jenkins.py", line 329, in delete_node
    del self.nodes[nodename]
  File "/tmp/ws/sesdev-integration/PR-351/8/jcs/venv/lib/python3.6/site-packages/jenkinsapi/nodes.py", line 125, in __delitem__
    raise UnknownNode('Node %s does not exist' % item)
jenkinsapi.custom_exceptions.UnknownNode: 'Node sesdev-integration_PR-351-8 does not exist'

which results in a non-deleted cloud instance which is bad.

+++ jcs/jen.py
@@ -33,8 +33,12 @@ class JenkinsClient:
 
     def delete_node(self, node_name):
         print('Jenkins node "{}" deleting...'.format(node_name))
-        self._client.delete_node(node_name)
-        print('Jenkins node "{}" deleted'.format(node_name))
+        nodes = self._client.get_nodes()
+        if node_name in nodes.keys():
+            self._client.delete_node(node_name)
+            print('Jenkins node "{}" deleted'.format(node_name))
+        else:
+            print('Jenkins node "{}" not found. Not deleted'.format(node_name))
 
     def create_node(self, node_hostname, node_name, node_desc, credential_desc,
                     node_labels, force=True):
